Add files via upload
Hack-AU  #hackathon at Microsoft was all about Social Entrepreneurship.

It was a pleasure to work and meet brilliant minds and ideas that can make our world a better place!

I am proud to share that we developed and presented a system, Fallkol, that will take care of the elderly population helping to identify and prevent their next fall-Gd forbid!

⭐About the system, FallKol. 
The system is an AI-based Acoustic Adult Fall Detection System. 

⭐How does it look and work?   In the -Front-End, the system can be installed on any technological device: smartphone, camera and/or basically any device that exists in older peoples' homes.
The system will work in the background and there will be no need to open the app or the code.

⭐What is happening in the Back-End?

As I mentioned the system is an AI and deep learning based system, 
which is based on a very large database which is divided to "fall" and "not fall" soundtracks.  It "learns" and "teaches" itself how an elderly falling person sounds and prevents from false alerts (False Positive).

👉Actually, there is so much more to say and explain and I'll be happy to share with anyone that is interested.